Step-04: Refactor and relocate `createFindAgentTool` to `FindAgent.kt`, simplify Langfuse observability setup
- Moved the `createFindAgentTool` function from `Main.kt` to `FindAgent.kt` for better cohesion.
- Simplified Langfuse exporter setup in `setupObservability`, removing environment variable checks.
